+++ webviz_config/webviz_plugin_subclasses/_views.py
@@ -8,15 +8,20 @@ import webviz_core_components as wcc
 
 from ._settings_group_abc import SettingsGroupABC
 
-from dash import Dash
+from dash import Dash, Input, Output
 from dash.development.base_component import Component
 
 
 class ViewElementABC(abc.ABC):
-    def __init__(self) -> None:
+    def __init__(
+        self, flex_grow: int = 1, screenshot_filename: str = "webviz-screenshot.png"
+    ) -> None:
         super().__init__()
 
+        self._flex_grow = flex_grow
         self._uuid: str = ""
+        self._screenshot_filename = screenshot_filename
+        self._add_download_button = False
 
         self._settings: List[SettingsGroupABC] = []
 
@@ -42,6 +47,15 @@ class ViewElementABC(abc.ABC):
 
         settings_group._set_uuid(uuid)
         self._settings.append(settings_group)
+
+    @property
+    def view_element_data_output(self) -> Output:
+        self._add_download_button = True
+        return Output(self.uuid(), "download")
+
+    @property
+    def view_element_data_requested(self) -> Input:
+        return Input(self.uuid(), "data_requested")
 
     def layout(self) -> Union[str, Type[Component]]:
         raise NotImplementedError
@@ -115,7 +129,12 @@ class ViewLayoutElement:
             return wcc.WebvizPluginLayoutRow(
                 flexGrow=self._flex_grow,
                 children=[
-                    wcc.WebvizViewElement(id=el.uuid(), children=el.layout())
+                    wcc.WebvizViewElement(
+                        id=el.uuid(),
+                        showDownload=el._add_download_button,
+                        flexGrow=el._flex_grow,
+                        children=el.layout(),
+                    )
                     if isinstance(el, ViewElementABC)
                     else el.layout()
                     for el in self._children
@@ -125,7 +144,12 @@ class ViewLayoutElement:
             return wcc.WebvizPluginLayoutColumn(
                 flexGrow=self._flex_grow,
                 children=[
-                    wcc.WebvizViewElement(id=el.uuid(), children=el.layout())
+                    wcc.WebvizViewElement(
+                        id=el.uuid(),
+                        showDownload=el._add_download_button,
+                        flexGrow=el._flex_grow,
+                        children=el.layout(),
+                    )
                     if isinstance(el, ViewElementABC)
                     else el.layout()
                     for el in self._children
@@ -253,7 +277,12 @@ class ViewABC(abc.ABC):
     def layout(self) -> Type[Component]:
         return html.Div(
             [
-                wcc.WebvizViewElement(id=el.uuid(), children=[el.layout()])
+                wcc.WebvizViewElement(
+                    id=el.uuid(),
+                    showDownload=el._add_download_button,
+                    flexGrow=el._flex_grow,
+                    children=[el.layout()],
+                )
                 if isinstance(el, ViewElementABC)
                 else el.layout
                 for el in self._layout_elements
